"PLFrontendOS::Linux_XDnDFileDropHelper::PropertyData": Fixed a minor cast issue (Clang under Mac OS X produced a compiler error in "PLFrontendOS::Linux_XDnDFileDropHelper::ReadProperty()")

+++ Plugins/PLFrontendOS/include/PLFrontendOS/Linux_XDnDFileDropHelper.h
@@ -97,7 +97,7 @@ class Linux_XDnDFileDropHelper {
 		*/
 		struct PropertyData {
 			unsigned char  *data;
-			unsigned long	format;
+			int				format;
 			unsigned long	nNumOfItems;
 			Atom			type;
 		};
